// libs/pos/src/lib/models/debt.models.ts
export interface Condition {
  field: string;
  operator: string;
  value: string;
  logicOperator: 'AND' | 'OR';
}